rh7: import RHEL7 kernel-3.10.0-1062.1.2.el7
https://rhn.redhat.com/errata/RHSA-2019-2829.html

https://jira.sw.ru/browse/PSBM-98031

+++ lib/checksum.c
@@ -182,9 +182,7 @@ EXPORT_SYMBOL(csum_partial_copy);
 
 #ifndef csum_tcpudp_nofold
 __wsum csum_tcpudp_nofold(__be32 saddr, __be32 daddr,
-			unsigned short len,
-			unsigned short proto,
-			__wsum sum)
+			  __u32 len, __u8 proto, __wsum sum)
 {
 	unsigned long long s = (__force u32)sum;
